Task Integration (#1)
* Testing CircleCI

* Testing new AWS IAM credentials/permissions

* Increased deployment timeout threshold (#64)

* Increased ecs_deploy timeout threshold to 180 seconds

The most recent deploy took about 120 seconds, but the current
threshold for the ecs_deploy script timeout is 90 seconds. This causes
builds on CircleCI to fail with the red X - not a good look. 180
seconds plays it on the safe side without being too long.

* Create task after classifer

* Task creation working

* Expand task on get

* Expand on classifier post

* Starting task creation tests

* Expanding task(s)

* Fixed tests

The unique ID of a task was causing problems in the testing environment
when integrating with a task-service container. This commit resolves
that problem by detecting if tests are running and generating a random
unique ID if that is the case.

* Forgot to import Gene

* Fixed task-def creation request data

In accordance with commits made in the task-service repository

* Added endpoint for completed notebook upload to classifier (for ml-worker)

When ml-worker completes running a notebook, it needs to upload the
completed notebook to core-service so that core-service can send an
email to the user with a link to download their completed notebook.

This commit enabled that functionality by adding:
- New authentication permission designed to only allow an internal
service to upload a notebook
- notebook_file attribute to Classifier model and serializer
- rudimentary file storage logic (stores files locally under
/media_files/notebook/classifier_<id>.ipynb, no S3 integration yet)
- Tests for notebook uploads, which include uploading a real notebook
- Whenever a test is ran that creates a file, you are always left with
the directory still on your filesystem after the test. I added a test
runner file which will delete the media_files directory after testing

* Make classifiers write-once only

For now, we will assume classifiers cannot be updated.