Consolidated task-service and core-service
All of the task-service functionality is now ported over into
core-service, including queueing, serialization, views, etc. All of the
relevant columns that used to be stored on Task and TaskDef objects in
task-service are now stored directly on the classifier in core-service.
This greatly simplifies Cognoma’s overall architecture and codebase.